main_branch/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-otp==1.3.0
django-user-agents==0.4.0
qrcode==7.4.2
pyotp==2.9.0